% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bed.R
\name{create_bed}
\alias{create_bed}
\title{Create BED, BIM and FAM for VCF}
\usage{
create_bed(filename, output_name, folder_name)
}
\arguments{
\item{filename}{The .vcf.gz file you want to create a PLINK .bed/.bim/.fam file for.}

\item{output_name}{Outputs the files output_name.bed, outputname.bim,
and outputname.fam}

\item{folder_name}{The name of the folder where your vcf file is located.}
}
\value{
A .bed, .bim, and .fam file all with the output file name (e.g., galaxy_inputs/genome.bed)
}
\description{
Uses the PLINK 1.9 library to create BED, BIM, and FAM
files from a VCF file without the hassle of typing in the command lines.
Includes the option to run in a remote Galaxy server with an RStudio instance.
}
\examples{
create_bed("genome.vcf.gz", "genome_output", folder_name="galaxy_inputs")


}
