% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxy_create_bed.R
\name{galaxy_create_bed}
\alias{galaxy_create_bed}
\title{Create BED, BIM and FAM for VCF}
\usage{
galaxy_create_bed(filename, output_name)
}
\arguments{
\item{filename}{The VCF file you want to create a PLINK .bed/.bim/.fam file for}

\item{output_name}{Outputs the files output_name.bed, outputname.bim,
and outputname.fam}
}
\value{
A .bed, .bim, and .fam file all with the output file name, found in the
galaxy_inputs directory (e.g., galaxy_inputs/genome.bed)
}
\description{
Uses the PLINK 1.9 library to create BED, BIM, and FAM
files from a VCF file.
}
\examples{
galaxy_create_bed("genome.vcf.gz", "genome_output")


}
