% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxy_create_pca.R
\name{galaxy_create_pca}
\alias{galaxy_create_pca}
\title{PCA for genetic populations}
\usage{
galaxy_create_pca(filename, n_components, output_name)
}
\arguments{
\item{filename}{The vcf.gz file prefix. Do not include any extensions.}

\item{n_components}{The number of dimensions you want to reduce the
genomic data to}

\item{output_name}{Output file name for the eigenvector and eigenvalue files}
}
\value{
A .eigenvec and .eigenval file with the output name.
}
\description{
Uses the PLINK 1.9 library to create an n-dimensional principal
components analysis for more effective clustering and data analysis.
}
\examples{
galaxy_create_pca("genome", 10, "genome_output")
galaxy_create_pca("chr1", 25, "chr1_output")


}
