% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxy_rename_file.R
\name{galaxy_rename_file}
\alias{galaxy_rename_file}
\title{Rename Galaxy file}
\usage{
galaxy_rename_file(oldName, newName)
}
\arguments{
\item{oldName}{The original file name}

\item{newName}{The new file name}
}
\value{
A new file within the galaxy_inputs directory.
}
\description{
Changes the file name within the Galaxy input directory
without needing to manually rename in the console.
}
\examples{
galaxy_rename_file("superduperlongfilenamethatnoonewillremember.vcf.gz", "genome.vcf.gz")
galaxy_rename_file("A109410_SNP_lifted_sorted_at.vcf.gz", "A109410_v2.vcf.gz")


}
