% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_euclidean_distance.R
\name{get_euclidean_distance}
\alias{get_euclidean_distance}
\title{Euclidean distance}
\usage{
get_euclidean_distance(vec1, vec2)
}
\arguments{
\item{vec1}{The first vector.}

\item{vec2}{The second vector.}
}
\value{
The euclidean distance between vec1 and vec2.
}
\description{
Calculates the euclidean distance between two numerical vectors
of the same length.
}
\examples{
# geographic example
ann_arbor <- c(42.27756, -83.74088)
glacier_bay <- c(58.6999972, -136.1499994)
get_euclidean_distance(ann_arbor, glacier_bay)
}
