% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_plink.R
\name{install_plink}
\alias{install_plink}
\title{Install PLINK}
\usage{
install_plink()
}
\description{
Downloads the PLINK 1.9 library on an RStudio
instance without the need to manually download through the
command line. This is especially useful for cloud-based servers,
such as Galaxy for Bioinformatics.
}
