% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxy_rename_file.R
\name{rename_file}
\alias{rename_file}
\title{Rename a file}
\usage{
rename_file(oldName, newName, folderName)
}
\arguments{
\item{oldName}{The original file name}

\item{newName}{The new file name}

\item{folder_name}{The name of the folder where your file is located.}
}
\value{
A new file within the galaxy_inputs directory.
}
\description{
Changes the file name within an input directory
without needing to manually rename in the console. This is
especially useful when working on a Galaxy RStudio instance.
}
\examples{
rename_file("superduperlongfilenamethatnoonewillremember.vcf.gz", "genome.vcf.gz", "galaxy_inputs")
rename_file("A109410_SNP_lifted_sorted_at.vcf.gz", "A109410_v2.vcf.gz", "myFolder")


}
