% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_euclidean_distance.R
\name{get_euclidean_distance}
\alias{get_euclidean_distance}
\title{Euclidean distance}
\usage{
get_euclidean_distance(vec1, vec2)
}
\arguments{
\item{vec1}{The first vector.}

\item{vec2}{The second vector.}
}
\value{
The euclidean distance between vec1 and vec2.
}
\description{
Calculates the euclidean distance between two numerical vectors
of the same length.
}
\examples{
get_euclidean_distance(c(1, 2, 3), c(4, 5, 6))
}
